Move custom attributes to the separate folder
