+++ src/ggrc-client/js/models/custom-attributes/index.js
@@ -1,0 +1,8 @@
+/*
+  Copyright (C) 2018 Google Inc.
+  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
+
+import './custom-attributable';
+import './custom-attribute-definition';
+import './custom-attribute-value';
